create visualization model including placeholder for dashboard
