+++ twitter_analysis/visualization/dashboard.py
@@ -1,0 +1,9 @@
+"""
+Placeholder module for future dashboard implementation.
+
+This could integrate temporal and content visualizations into an interactive interface
+using tools like Streamlit or Dash.
+"""
+
+def create_dashboard() -> None:
+    """Future function to create an interactive dashboard."""
